Move CoroutineTestUtils to unit test package
